refactor: 도메인 패키지를 car와 game으로 분리
