+++ src/main/java/racingcar/domain/game/Lap.java
@@ -1,4 +1,4 @@
-package racingcar.domain;
+package racingcar.domain.game;
 
 public class Lap {
