+++ src/main/java/racingcar/domain/game/MathNumberGenerator.java
@@ -1,4 +1,4 @@
-package racingcar.domain;
+package racingcar.domain.game;
 
 public class MathNumberGenerator implements NumberGenerator {
