+++ src/main/java/racingcar/step/CreateCarStep.java
@@ -2,7 +2,7 @@ package racingcar.step;
 
 import racingcar.common.log.Logger;
 import racingcar.controller.RacingCarController;
-import racingcar.domain.Cars;
+import racingcar.domain.car.Cars;
 import racingcar.step.context.GameContext;
 import racingcar.view.InputView;
