+++ src/test/java/racingcar/domain/car/CarsTest.java
@@ -1,9 +1,11 @@
-package racingcar.domain;
+package racingcar.domain.car;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.provider.Arguments;
+import racingcar.mock.MockFixedNumberGenerator;
+import racingcar.util.WinnerCarsHelper;
 
 import java.util.List;
 import java.util.stream.Collectors;
