+++ src/test/java/racingcar/domain/game/MathNumberGeneratorTest.java
@@ -1,4 +1,4 @@
-package racingcar.domain;
+package racingcar.domain.game;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Disabled;
